Move robust math functions to a common module.
